refactor: layer input handling as a pure decision core + Editor shell
app/input.rs was 3,584 lines of impl Editor mixing four concerns: key
routing, action execution, overlay focus queries, and agent-script
evaluation. Decisions about what a key *means* were interleaved with
the state mutations that carry them out, so none of that logic could
be exercised without constructing an editor.

Split it into an Editor-free decision core and a high-level shell:

- input/router.rs (new, low level): pure functions that never see
  Editor. Each takes only the narrow inputs it needs — the key event,
  the KeybindingResolver, a small read-only view struct — and returns
  a decision value: layout_reading (which reading of a chord wins),
  widget_panel_key -> WidgetKeyOutcome (the whole floating-panel/dock
  decision tree), mode_key_disposition -> ModeKeyDisposition (mode
  chords/bindings, text-input capture, clipboard forwarding, Shift+nav
  widget selection, read-only blocking), chord_or_key -> ChordDisposition,
  the popup key-precedence resolvers, should_dismiss_transient_popup,
  and cancels_pending_lsp. Being pure, they carry 14 unit tests that
  run without an editor.

- app/input.rs (high level): the imperative shell. handle_key owns the
  pipeline order and applies effects — build the view, ask the router
  what the key means, execute the outcome it names.

- app/action_dispatch.rs: Editor::handle_action. Executing a resolved
  Action drives buffers, windows, prompts and plugins, so it stays a
  method; the split keeps deciding what a key means separate from
  doing what the action says.

- app/overlay.rs: gains the overlay-stack builder and the shared focus
  queries, plus a pure popup_blocked_by_higher_modal beside the
  existing layer predicates — extending the pattern that file already
  established.

- app/agent_scripts.rs: eval_agent_script, which was never input
  handling.

- Window::claim_next_key: the plugin getNextKey queue/buffer operation
  moves onto Window, where that state lives.

Behaviour-preserving. Two details are deliberate: chord state is
cleared only when a mode actually ran its resolution, so Normal-context
chord prefixes survive the mode stage; and ChordDisposition separates
Chord from Resolved because a completed chord never cancelled in-flight
LSP requests. The router's tests pin the default keymap so they assert
router semantics rather than host-OS defaults (Config::default() picks
the macos keymap on macOS, where Ctrl+A is MoveLineStart, not SelectAll).
